move WriteCloneIntoRaw into alloc::alloc
